--- a/openh264/src/decoder.rs
+++ b/openh264/src/decoder.rs
@@ -50,7 +50,6 @@
 
 use crate::error::NativeErrorExt;
 use crate::formats::yuv2rgb::{write_rgb8_f32x8_par, write_rgba8_f32x8, write_rgba8_scalar};
-// use crate::formats::yuv2rgb::{write_rgb8_f32x8, write_rgb8_f32x8_par, write_rgb8_scalar, write_rgb8_scalar, write_rgba8_f32x8, write_rgba8_scalar_par};
 use crate::formats::{YUVSlices, YUVSource};
 use crate::{Error, OpenH264API, Timestamp};
 use openh264_sys2::{