--- a/openh264/src/decoder.rs
+++ b/openh264/src/decoder.rs
@@ -49,7 +49,7 @@
 //! ```
 
 use crate::error::NativeErrorExt;
-use crate::formats::yuv2rgb::{write_rgb8_f32x8_par, write_rgba8_f32x8, write_rgba8_scalar};
+use crate::formats::yuv2rgb::{write_rgb8_f32x8, write_rgb8_scalar, write_rgba8_f32x8, write_rgba8_scalar};
 use crate::formats::{YUVSlices, YUVSource};
 use crate::{Error, OpenH264API, Timestamp};
 use openh264_sys2::{
@@ -571,15 +571,14 @@ impl DecodedYUV<'_> {
             target.len()
         );
 
-        write_rgb8_f32x8_par(self.y, self.u, self.v, dim, strides, target);
-        // // for f32x8 math, image needs to:
-        // //   - have a width divisible by 8
-        // //   - have at least two rows
-        // if dim.0 % 8 == 0 && dim.1 >= 2 {
-        //     write_rgb8_f32x8(self.y, self.u, self.v, dim, strides, target);
-        // } else {
-        //     write_rgb8_scalar(self.y, self.u, self.v, dim, strides, target);
-        // }
+        // for f32x8 math, image needs to:
+        //   - have a width divisible by 8
+        //   - have at least two rows
+        if dim.0 % 8 == 0 && dim.1 >= 2 {
+            write_rgb8_f32x8(self.y, self.u, self.v, dim, strides, target);
+        } else {
+            write_rgb8_scalar(self.y, self.u, self.v, dim, strides, target);
+        }
     }
 
     // TODO: Ideally we'd like to move these out into a converter in `formats`.