--- a/openh264/src/formats/yuv2rgb.rs
+++ b/openh264/src/formats/yuv2rgb.rs
@@ -434,7 +434,9 @@ fn write_rgba8_f32x8_row(y_row: &[u8], u_row: &[u8], v_row: &[u8], width: usize,
 #[cfg(test)]
 mod test {
     use crate::decoder::{Decoder, DecoderConfig};
-    use crate::formats::yuv2rgb::{write_rgb8_f32x8, write_rgb8_f32x8_par, write_rgb8_scalar, write_rgb8_scalar_par, write_rgba8_scalar};
+    use crate::formats::yuv2rgb::{
+        write_rgb8_f32x8, write_rgb8_f32x8_par, write_rgb8_scalar, write_rgb8_scalar_par, write_rgba8_scalar,
+    };
     use crate::formats::YUVSource;
     use crate::OpenH264API;