This extends the effort from the previous two PRs.
Two methods have been added, which implement a parallel approach to  `write_rgb8_scalar` and `write_rgb8_f32x8`, without rayon ;)


**f32x8**
```
Before
test convert_yuv_to_rgb_1920x1080  ... bench:   4,425,583.45 ns/iter (+/- 471,817.41)
test convert_yuv_to_rgb_512x512    ... bench:     588,151.05 ns/iter (+/- 144,696.95)

After
test convert_yuv_to_rgb_1920x1080  ... bench:   1,291,415.29 ns/iter (+/- 182,767.64)
test convert_yuv_to_rgb_512x512    ... bench:     241,422.93 ns/iter (+/- 19,608.33)
```


**scalar**
```
Before
test convert_yuv_to_rgb_1920x1080  ... bench:   8,984,152.55 ns/iter (+/- 1,119,268.56)
test convert_yuv_to_rgb_512x512    ... bench:   1,122,767.55 ns/iter (+/- 62,880.50)

After
test convert_yuv_to_rgb_1920x1080  ... bench:   3,457,987.00 ns/iter (+/- 2,434,315.60)
test convert_yuv_to_rgb_512x512    ... bench:     444,836.04 ns/iter (+/- 315,727.15)
```

Please inspect