@jnnks 2025-04-27
> _par should not be the default, ...

Ok, I defaulted to `write_rgb8_f32x8_par` just for the benchmarks.


I like the idea with `.split_n()`. One of the issues I have with this impl is the weird offset calculations (eg. `let base_y = (2 * (y + offset) + 1) * strides.0;`). We can clean that up nicely and concentrate the index math into one place.