@jnnks 2025-05-03
First impl for `DecodedYUV::split`. 

Some thoughts:
* I used  `split<const N: usize>` to return an array instead of an iterator. This feels a bit more in line with the idea of using this for multi threading, where the number of threads would be const as well.
* Internally this is implemented with `chunks` iterators and the array is built from those iterators. This can certainly be improved later on. The last chunk will be larger than the others if the buffer cannot be split evenly.
* `split` returns `(&[u8], &[u8], &[u8])`. I suppose `YUVSlices` would be better?
* I am confused about `DecodedYUV`, `YUVSource` and `YUVBuffer`. Where should this impl live?