@ralfbiedert 2025-05-23
Hey, so sorry, I somehow missed your latest update!

> I used split<const N: usize> to return an array instead of an iterator. This feels a bit more in line with the idea of using this for multi threading, where the number of threads would be const as well.

Personally, I don't mind this being a const parameter, but I'd assume some folks would like this to be configurable (e.g., detect numcpus). For now const is fine though.

> split returns (&[u8], &[u8], &[u8]). I suppose YUVSlices would be better?

Yes, I think each one being `YUVSlices` would be nicer to use.

> I am confused about DecodedYUV, YUVSource and YUVBuffer. Where should this impl live?

You mean where `split` should live? To be honest I'd have to play with this a bit myself, but we can refactor that later. 

Let me know if there's anything I can help with.