@jnnks 2025-06-09
I am using `clippy 0.1.87 (17067e9ac6 2025-05-09)` and there are other warnings. The one the CI complained about is not one of those:

```
error: unnecessary semicolon
   --> openh264/src/decoder.rs:391:14
    |
391 |             };
    |              ^ help: remove
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_semicolon
    = note: `-D clippy::unnecessary-semicolon` implied by `-D clippy::pedantic`
    = help: to override `-D clippy::pedantic` add `#[allow(clippy::unnecessary_semicolon)]`

error: this could be a `const fn`
   --> openh264/src/decoder.rs:423:5
    |
423 | /     pub unsafe fn raw_api(&mut self) -> &mut DecoderRawAPI {
424 | |         &mut self.raw_api
425 | |     }
    | |_____^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#missing_const_for_fn
    = note: `-D clippy::missing-const-for-fn` implied by `-D clippy::nursery`
    = help: to override `-D clippy::nursery` add `#[allow(clippy::missing_const_for_fn)]`
help: make the function `const`
    |
423 |     pub const unsafe fn raw_api(&mut self) -> &mut DecoderRawAPI {
    |         +++++

error: this could be a `const fn`
   --> openh264/src/encoder.rs:810:5
    |
810 | /     pub unsafe fn raw_api(&mut self) -> &mut EncoderRawAPI {
811 | |         &mut self.raw_api
812 | |     }
    | |_____^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#missing_const_for_fn
help: make the function `const`
    |
810 |     pub const unsafe fn raw_api(&mut self) -> &mut EncoderRawAPI {
    |         +++++

error: unnecessary semicolon
   --> openh264/src/encoder.rs:896:18
    |
896 |                 };
    |                  ^ help: remove
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_semicolon
```