@ralfbiedert 2025-06-10
> I am using clippy 0.1.87 (17067e9ac6 2025-05-09) and there are other warnings. The one the CI complained about is not one of those:

These should be fixed on the latest master. If you could rebase clippy should pass. If not (or in any case) we probably want to upgrade MSRV slowly anyways as more items become `const`.