@jnnks 2025-06-15
> If you could rebase ...

I merged instead with an embarrassing amount of push --force. Lmk if you want rebase instead

> I'm a bit confused, OpenH264 is only YUV420; so in other words, our own YUV conversion should only ever consider YUV420, or am I missing something?

Ah, sorry, my bad. I messed the dimensions in split()