@jnnks 2025-06-15
I am wondering how to continue here.

We can introduce a trait (eg. IntoRGB8) that can make the conversion available and we impl it for  DecodedYUV and YUVSlices. For YUVSlices this would be a bit awkward, because we need to offset the target reference outside of the function.

openh264/src/formats/rgb.rs already contains some RGBSlices. Conversion between YUV/RGBSlice would be independent from other parts of the frame and can be parallelized in any flavor. We would then have to join the slices somehow.

Seems to me that YUV/RGBSlices are not used in this repo, but part of the public API.
How are you using them?
From my POV, the second option makes more sense.